Extract row manipulation on Cassandra message table
