+++ mailbox/cassandra/src/main/java/org/apache/james/mailbox/cassandra/mail/CassandraMessageMapper.java
@@ -25,21 +25,11 @@ import static com.datastax.driver.core.querybuilder.QueryBuilder.incr;
 import static com.datastax.driver.core.querybuilder.QueryBuilder.select;
 import static com.datastax.driver.core.querybuilder.QueryBuilder.update;
 import static org.apache.james.mailbox.cassandra.table.CassandraMessageIds.MESSAGE_ID;
-import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.ATTACHMENTS;
-import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.BODY_CONTENT;
-import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.BODY_START_OCTET;
-import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.FULL_CONTENT_OCTETS;
-import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.HEADER_CONTENT;
-import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.INTERNAL_DATE;
-import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.MOD_SEQ;
-import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.PROPERTIES;
-import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.TEXTUAL_LINE_COUNT;
 import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.Flag.DELETED;
 import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.Flag.RECENT;
 import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.Flag.SEEN;
 
 import java.util.Comparator;
-import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
@@ -48,7 +38,6 @@ import java.util.stream.Collectors;
 
 import javax.mail.Flags;
 import javax.mail.Flags.Flag;
-import javax.mail.util.SharedByteArrayInputStream;
 
 import org.apache.james.backends.cassandra.utils.CassandraUtils;
 import org.apache.james.backends.cassandra.utils.FunctionRunnerWithRetry;
@@ -59,15 +48,8 @@ import org.apache.james.mailbox.cassandra.CassandraId;
 import org.apache.james.mailbox.cassandra.CassandraMessageId;
 import org.apache.james.mailbox.cassandra.mail.utils.MessageDeletedDuringFlagsUpdateException;
 import org.apache.james.mailbox.cassandra.table.CassandraMailboxCountersTable;
-import org.apache.james.mailbox.cassandra.table.CassandraMessageTable;
-import org.apache.james.mailbox.cassandra.table.CassandraMessageTable.Attachments;
-import org.apache.james.mailbox.cassandra.table.CassandraMessageTable.Properties;
 import org.apache.james.mailbox.exception.MailboxException;
-import org.apache.james.mailbox.model.Attachment;
-import org.apache.james.mailbox.model.AttachmentId;
-import org.apache.james.mailbox.model.Cid;
 import org.apache.james.mailbox.model.ComposedMessageId;
-import org.apache.james.mailbox.model.MessageAttachment;
 import org.apache.james.mailbox.model.MessageMetaData;
 import org.apache.james.mailbox.model.MessageRange;
 import org.apache.james.mailbox.model.UpdatedFlags;
@@ -79,20 +61,13 @@ import org.apache.james.mailbox.store.mail.ModSeqProvider;
 import org.apache.james.mailbox.store.mail.UidProvider;
 import org.apache.james.mailbox.store.mail.model.Mailbox;
 import org.apache.james.mailbox.store.mail.model.MailboxMessage;
-import org.apache.james.mailbox.store.mail.model.impl.PropertyBuilder;
-import org.apache.james.mailbox.store.mail.model.impl.SimpleMailboxMessage;
-import org.apache.james.mailbox.store.mail.model.impl.SimpleProperty;
 
 import com.datastax.driver.core.ResultSet;
 import com.datastax.driver.core.Row;
 import com.datastax.driver.core.Session;
-import com.datastax.driver.core.UDTValue;
 import com.datastax.driver.core.querybuilder.Assignment;
-import com.github.fge.lambdas.Throwing;
-import com.github.steveash.guavate.Guavate;
 import com.google.common.base.Throwables;
 import com.google.common.collect.ImmutableList;
-import com.google.common.primitives.Bytes;
 
 public class CassandraMessageMapper implements MessageMapper {
 
@@ -293,20 +268,12 @@ public class CassandraMessageMapper implements MessageMapper {
         try {
             ComposedMessageId messageId = retrieveComposedMessageId(CassandraMessageId.of(row.getUUID(MESSAGE_ID)));
 
-            SimpleMailboxMessage message =
-                    new SimpleMailboxMessage(
-                            messageId.getMessageId(),
-                            row.getDate(INTERNAL_DATE),
-                            row.getLong(FULL_CONTENT_OCTETS),
-                            row.getInt(BODY_START_OCTET),
-                            buildContent(row, fetchType),
-                            getFlags(row),
-                            getPropertyBuilder(row),
-                            messageId.getMailboxId(),
-                            getAttachments(row, fetchType));
-            message.setUid(messageId.getUid());
-            message.setModSeq(row.getLong(MOD_SEQ));
-            return message;
+            return CassandraMessageRowHandler.builder()
+                        .row(row)
+                        .messageId(messageId)
+                        .loadingAttachmentsFunction(attachmentMapper::getAttachments)
+                        .build()
+                    .toMailboxMessage(fetchType);
         } catch (MailboxException e) {
             throw Throwables.propagate(e);
         }
@@ -316,67 +283,6 @@ public class CassandraMessageMapper implements MessageMapper {
         return imapUidDAO.retrieve(messageId, Optional.empty())
             .findFirst()
             .orElseThrow(() -> new MailboxException("Message not found: " + messageId));
-    }
-
-    private Flags getFlags(Row row) {
-        Flags flags = new Flags();
-        for (String flag : CassandraMessageTable.Flag.ALL) {
-            if (row.getBool(flag)) {
-                flags.add(CassandraMessageTable.Flag.JAVAX_MAIL_FLAG.get(flag));
-            }
-        }
-        row.getSet(CassandraMessageTable.Flag.USER_FLAGS, String.class)
-            .stream()
-            .forEach(flags::add);
-        return flags;
-    }
-
-    private PropertyBuilder getPropertyBuilder(Row row) {
-        PropertyBuilder property = new PropertyBuilder(
-            row.getList(PROPERTIES, UDTValue.class).stream()
-                .map(x -> new SimpleProperty(x.getString(Properties.NAMESPACE), x.getString(Properties.NAME), x.getString(Properties.VALUE)))
-                .collect(Collectors.toList()));
-        property.setTextualLineCount(row.getLong(TEXTUAL_LINE_COUNT));
-        return property;
-    }
-
-    private List<MessageAttachment> getAttachments(Row row, FetchType fetchType) {
-        switch (fetchType) {
-        case Full:
-        case Body:
-            List<UDTValue> udtValues = row.getList(ATTACHMENTS, UDTValue.class);
-            Map<AttachmentId,Attachment> attachmentsById = attachmentsById(row, udtValues);
-
-            return udtValues
-                    .stream()
-                    .map(Throwing.function(x -> 
-                        MessageAttachment.builder()
-                            .attachment(attachmentsById.get(attachmentIdFrom(x)))
-                            .name(x.getString(Attachments.NAME))
-                            .cid(com.google.common.base.Optional.fromNullable(x.getString(Attachments.CID)).transform(Cid::from))
-                            .isInline(x.getBool(Attachments.IS_INLINE))
-                            .build()))
-                    .collect(Guavate.toImmutableList());
-        default:
-            return ImmutableList.of();
-        }
-    }
-
-    private Map<AttachmentId,Attachment> attachmentsById(Row row, List<UDTValue> udtValues) {
-        Map<AttachmentId, Attachment> map = new HashMap<>();
-        attachmentMapper.getAttachments(attachmentIds(udtValues)).stream()
-                .forEach(att -> map.put(att.getAttachmentId(), att));
-        return map;
-    }
-
-    private List<AttachmentId> attachmentIds(List<UDTValue> udtValues) {
-        return udtValues.stream()
-            .map(this::attachmentIdFrom)
-            .collect(Guavate.toImmutableList());
-    }
-
-    private AttachmentId attachmentIdFrom(UDTValue udtValue) {
-        return AttachmentId.from(udtValue.getString(Attachments.ID));
     }
 
     private MessageMetaData save(Mailbox mailbox, MailboxMessage message) throws MailboxException {
@@ -445,35 +351,6 @@ public class CassandraMessageMapper implements MessageMapper {
             message(Optional.ofNullable(messageDAO.selectMessageData(ImmutableList.of(messageId), FetchType.Metadata).join().one())
                 .orElseThrow(() -> new MessageDeletedDuringFlagsUpdateException(mailboxId, messageId)),
                 FetchType.Metadata));
-    }
-
-    private SharedByteArrayInputStream buildContent(Row row, FetchType fetchType) {
-        switch (fetchType) {
-            case Full:
-                return new SharedByteArrayInputStream(getFullContent(row));
-            case Headers:
-                return new SharedByteArrayInputStream(getFieldContent(HEADER_CONTENT, row));
-            case Body:
-                return new SharedByteArrayInputStream(getBodyContent(row));
-            case Metadata:
-                return new SharedByteArrayInputStream(new byte[]{});
-            default:
-                throw new RuntimeException("Unknown FetchType " + fetchType);
-        }
-    }
-
-    private byte[] getFullContent(Row row) {
-        return Bytes.concat(getFieldContent(HEADER_CONTENT, row), getFieldContent(BODY_CONTENT, row));
-    }
-
-    private byte[] getBodyContent(Row row) {
-        return Bytes.concat(new byte[row.getInt(BODY_START_OCTET)], getFieldContent(BODY_CONTENT, row));
-    }
-
-    private byte[] getFieldContent(String field, Row row) {
-        byte[] headerContent = new byte[row.getBytes(field).remaining()];
-        row.getBytes(field).get(headerContent);
-        return headerContent;
     }
 
 }
